Include aliases in the Available operations
The Available abstract operations (e.g. AvailableCalendars) should return
all possible aliases, so that other places in the spec (e.g. the
Temporal.Calendar constructor) can use them to determine whether a given
input value is valid. This input value can subsequently be canonicalized
by another abstract operation (e.g. CanonicalizeCalendar).

In Intl.supportedValuesOf(), on the other hand, we should _not_ return all
possible aliases, so we filter them out using a Canonicalize operation
before returning the list of Available codes as an array to the caller.

Not all of the kinds of codes here have aliases, and not even all of them
have a concept of "canonical". A quick investigation shows:

- Calendar: aliased; case-regularized; limited values "available" but any
  well-formed value accepted, unknown values coerced to the locale's
  default

- Collation: aliased, but web reality is that aliases are ignored;
  case-regularized; limited values "available" but any well-formed value
  accepted, unknown values coerced to the locale's default

- Currency: not sure if it is aliased because I don't have a copy of
  ISO 4217; case-regularized; limited values "available" but any
  well-formed value accepted and used

- Numbering system: not currently aliased; not case-regularized; limited
  values "available" but any well-formed value accepted, unknown values
  coerced to the locale's default

- Time zone: aliased; case-regularized

- Unit of measurement: not aliased; not case-regularized; limited values
  "available" but simple combinations of core values also accepted and
  used

So, I conclude that we need Canonicalize operations for calendars, time
zones, collation types, numbering systems, and possibly currency units.

An alternative approach would be to write Canonicalize operations for all
of the kinds of codes, and have them perform the case-regularization (and
for units of measurement they would be no-ops).

Closes: #37